requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlmodel>=0.0.16
sqlalchemy>=2.0
psycopg[binary]>=3.1
requests>=2.31
fecfile>=0.4
jinja2>=3.1
python-dateutil>=2.9
gunicorn>=21.2
python-dotenv
